backend/requirements.txt
openvino_genai
openvino
pillow # PIL
numpy
flask
flask-cors